move input_events.strategy to a base one
